Consolidate hasDynamicSignalTypes eval
