+++ x-pack/platform/plugins/shared/fleet/common/services/otelcol_helpers.test.ts
@@ -7,11 +7,7 @@
 
 import type { PackageInfo } from '../types';
 
-import {
-  packageInfoHasOtelInputs,
-  hasDynamicSignalTypes,
-  packagePolicyHasOtelInputs,
-} from './otelcol_helpers';
+import { packageInfoHasOtelInputs, packagePolicyHasOtelInputs } from './otelcol_helpers';
 
 describe('packageInfoHasOtelInputs', () => {
   const otelPackageInfo = {
@@ -67,161 +63,6 @@ describe('packageInfoHasOtelInputs', () => {
 
   it('should return false if undefined is passed', () => {
     expect(packageInfoHasOtelInputs(undefined)).toBe(false);
-  });
-});
-
-describe('hasDynamicSignalTypes', () => {
-  it('should return true when OTel input has dynamic_signal_types: true', () => {
-    const pkg = {
-      policy_templates: [
-        {
-          name: 't',
-          title: 't',
-          input: 'otelcol',
-          type: 'logs',
-          template_path: 'input.yml.hbs',
-          dynamic_signal_types: true,
-        },
-      ],
-    } as any as PackageInfo;
-    expect(hasDynamicSignalTypes(pkg)).toBe(true);
-  });
-
-  it('should return false when OTel input has dynamic_signal_types: false', () => {
-    const pkg = {
-      policy_templates: [
-        {
-          name: 't',
-          title: 't',
-          input: 'otelcol',
-          type: 'logs',
-          template_path: 'input.yml.hbs',
-          dynamic_signal_types: false,
-        },
-      ],
-    } as any as PackageInfo;
-    expect(hasDynamicSignalTypes(pkg)).toBe(false);
-  });
-
-  it('should return false when OTel input has no dynamic_signal_types property', () => {
-    const pkg = {
-      policy_templates: [
-        { name: 't', title: 't', input: 'otelcol', type: 'logs', template_path: 'input.yml.hbs' },
-      ],
-    } as any as PackageInfo;
-    expect(hasDynamicSignalTypes(pkg)).toBe(false);
-  });
-
-  it('should return false when policyTemplateName does not match any template', () => {
-    const pkg = {
-      policy_templates: [
-        {
-          name: 'otel_template',
-          title: 't',
-          input: 'otelcol',
-          type: 'logs',
-          template_path: 'input.yml.hbs',
-          dynamic_signal_types: true,
-        },
-      ],
-    } as any as PackageInfo;
-    expect(hasDynamicSignalTypes(pkg, { policyTemplateName: 'other_template' })).toBe(false);
-  });
-
-  it('should return true when scoped to the matching policyTemplateName', () => {
-    const pkg = {
-      policy_templates: [
-        {
-          name: 'otel_template',
-          title: 't',
-          input: 'otelcol',
-          type: 'logs',
-          template_path: 'input.yml.hbs',
-          dynamic_signal_types: true,
-        },
-      ],
-    } as any as PackageInfo;
-    expect(hasDynamicSignalTypes(pkg, { policyTemplateName: 'otel_template' })).toBe(true);
-  });
-
-  it('should return false when packageInfo is undefined', () => {
-    expect(hasDynamicSignalTypes(undefined)).toBe(false);
-  });
-
-  it('should return true for a composable template where any input has dynamic_signal_types', () => {
-    const pkg = {
-      policy_templates: [
-        {
-          name: 'my_template',
-          title: 't',
-          inputs: [
-            { type: 'otelcol', title: 't', dynamic_signal_types: true },
-            { type: 'logfile', title: 't' },
-          ],
-        },
-      ],
-    } as any as PackageInfo;
-    expect(hasDynamicSignalTypes(pkg)).toBe(true);
-  });
-
-  it('should return true for a composable template scoped by inputType to the dynamic input', () => {
-    const pkg = {
-      policy_templates: [
-        {
-          name: 'my_template',
-          title: 't',
-          inputs: [
-            { type: 'otelcol', title: 't', dynamic_signal_types: true },
-            { type: 'logfile', title: 't' },
-          ],
-        },
-      ],
-    } as any as PackageInfo;
-    expect(
-      hasDynamicSignalTypes(pkg, { policyTemplateName: 'my_template', inputType: 'otelcol' })
-    ).toBe(true);
-  });
-
-  it('should return false for a composable template scoped by inputType to a non-dynamic input', () => {
-    const pkg = {
-      policy_templates: [
-        {
-          name: 'my_template',
-          title: 't',
-          inputs: [
-            { type: 'otelcol', title: 't', dynamic_signal_types: true },
-            { type: 'logfile', title: 't' },
-          ],
-        },
-      ],
-    } as any as PackageInfo;
-    expect(
-      hasDynamicSignalTypes(pkg, { policyTemplateName: 'my_template', inputType: 'logfile' })
-    ).toBe(false);
-  });
-
-  it('should scope to the correct template in a multi-template package', () => {
-    const pkg = {
-      policy_templates: [
-        {
-          name: 'otel_template',
-          title: 't',
-          input: 'otelcol',
-          type: 'logs',
-          template_path: 'input.yml.hbs',
-          dynamic_signal_types: true,
-        },
-        {
-          name: 'logfile_template',
-          title: 't',
-          input: 'logfile',
-          type: 'logs',
-          template_path: 'input.yml.hbs',
-        },
-      ],
-    } as any as PackageInfo;
-    expect(hasDynamicSignalTypes(pkg, { policyTemplateName: 'otel_template' })).toBe(true);
-    expect(hasDynamicSignalTypes(pkg, { policyTemplateName: 'logfile_template' })).toBe(false);
   });
 });
