+++ x-pack/platform/plugins/shared/fleet/common/services/policy_template.ts
@@ -131,6 +131,29 @@ export function registryInputAllowsDynamicSignalTypes(input: RegistryInput): boo
 }
 
 /**
+ * Returns true when any policy template in the package contains an input that
+ * declares dynamic signal types (dynamic_signal_types: true).
+ *
+ * Optionally, you can scope the query to a specific policy template and/or input type.
+ *
+ * Covers both:
+ *   - Input-only packages (top-level `input` key on the policy template)
+ *   - Composable integration packages (nested `inputs[]` entries)
+ */
+export const hasDynamicSignalTypes = (
+  packageInfo: PackageInfo | undefined,
+  scope?: { policyTemplateName?: string; inputType?: string }
+): boolean =>
+  (packageInfo?.policy_templates ?? []).some((template) => {
+    if (scope?.policyTemplateName && template.name !== scope.policyTemplateName) return false;
+    const inputs = getNormalizedInputs(template);
+    const relevant = scope?.inputType
+      ? inputs.filter((input) => input.type === scope.inputType)
+      : inputs;
+    return relevant.some(registryInputAllowsDynamicSignalTypes);
+  });
+
+/**
  * Returns true when the given package policy input corresponds to a registry input
  * that allows undefined data_stream.type (i.e. dynamic_signal_types).
  *
@@ -142,27 +165,10 @@ export function packagePolicyInputAllowsUndefinedDataStreamType(
   packageInfo: PackageInfo,
   packagePolicyInput: Pick<NewPackagePolicyInput | PackagePolicyInput, 'type' | 'policy_template'>
 ): boolean {
-  const templates = packageInfo.policy_templates ?? [];
-  for (const template of templates) {
-    if (isInputOnlyPolicyTemplate(template)) {
-      if (template.input === packagePolicyInput.type) {
-        const def = getPolicyTemplateInputDefinition(template);
-        return def ? registryInputAllowsDynamicSignalTypes(def) : false;
-      }
-    } else {
-      if (
-        packagePolicyInput.policy_template &&
-        template.name !== packagePolicyInput.policy_template
-      ) {
-        continue;
-      }
-      const def = getPolicyTemplateInputDefinition(template, packagePolicyInput.type);
-      if (def) {
-        return registryInputAllowsDynamicSignalTypes(def);
-      }
-    }
-  }
-  return false;
+  return hasDynamicSignalTypes(packageInfo, {
+    policyTemplateName: packagePolicyInput.policy_template,
+    inputType: packagePolicyInput.type,
+  });
 }
 
 export function getNormalizedDataStreams(
