+++ x-pack/platform/plugins/shared/fleet/public/applications/fleet/sections/agent_policy/create_package_policy_page/components/steps/components/package_policy_input_stream.tsx
@@ -43,8 +43,8 @@ import {
   mapPackageReleaseToIntegrationCardRelease,
   DATA_STREAM_USE_APM_VAR,
   shouldIncludeUseAPMVar,
+  hasDynamicSignalTypes,
 } from '../../../../../../../../../common/services';
-import { hasDynamicSignalTypes } from '../../../../../../../../../common/services/otelcol_helpers';
 
 import type {
   NewPackagePolicyInputStream,
